refactor(uzicli): split review+findings verbs into client_review.go/fake_review.go (PRD #1017 M2)
Pure motion: move the review/findings HTTPClient verbs (with dispositionPath and
the two review-filing types) and their FakeClient twins into per-domain files.
go doc -u -all byte-identical; no test edits.
